feat: add support for non-derived one-to-many relations
Currently you can define one-to-many relations using @derivedFrom
directive. Using it the field on the entity is just virtual field
that gets filled in at query time using lookup on the derived entity's
column.

This PR adds support for non-derived one-to-many relations -
this field needs to be filled in by writers to be usable.
You specify ID's of referenced entity and those will be filled in
with full object at query time.