fix: only optimistically resolve nested entities
We have this optimization that nested entities will resolve id if it's
only requested field without making actual database query.

This might not be ideal, if relation is broken (for example Space references
not existing SpaceMetadata), but can save some queries, so it's mostly
just a convention.

However this ended up also short-circuting requests like this:
```gql
query {
  space(id: "eth:doesnotexist") {
    id
  }
}
```

This commit prevents this from happening, by checking if it was resolved
from parent.