fix: read type on-demand (#326)
* fix: read type on-demand

We can't precompute type because it's impossible to easily figure out
what part of where filter is type and what is operator.

For example max_end_gte should be (max_end, gte), but authenticators_not_containts
should be (authenticators, not_contains) so it can't be done with simple split,
because pivot is moving around.

* chore: bump version to 0.1.0-beta.44