fix: fetch missing logs if new source is added (#370)
* chore: remove seenPoolTransactions in EVM provider

* refactor: remove txIndex from handleTx

txIndex is no longer useful as it's not exactly txIndex in the block
as we don't have full blocks in some cases

* refactor: delegate handleNewSourceAdded to provider only

* refactor: remove eventIndex from writer params

eventIndex is not stable similar to txIndex.

On EVM it can be accessed from Log directly.

* fix: fetch missing logs on new sources

Currently we don't handle a case where we process block/tx and new
events should be handled (due to new source being added).

This PR fixes it by fetching missing logs/events.